adjustments, fixes, health bar, and importing assets
So i worked on making the environment more better, also the feel of the game more better. Worked on implementing a health system that just feels right. Create boss figure 1, with a melee attack but their mechanic is still in the works, second figure is being created. I added more assets that could help make the game feel better or use the assets to create animations for the weapons they wield. I will have the movement of the boss working and some way to avoid him and heal.